# src/main/java/com/example/category/controller/dto/BookCategoriesDto.java
package com.example.category.controller.dto;

import java.util.List;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class BookCategoriesDto {

    private List<BookCategoryDto> categories;
}