# src/main/java/com/example/category/controller/dto/CreateBookCategoryDto.java
package com.example.category.controller.dto;

import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class CreateBookCategoryDto {

    @NotNull
    private String name;
}