# src/main/java/com/example/customer/controller/dto/CustomersDto.java
package com.example.customer.controller.dto;

import java.util.List;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class CustomersDto {

    private List<CustomerDto> customers;
}